Fix browser jsonptransport error
