+++ browser/lib/util/http.ts
@@ -9,6 +9,8 @@ import Realtime from '../../../common/lib/client/realtime';
 import XHRRequest from '../transport/xhrrequest';
 import XHRStates from '../../../common/constants/XHRStates';
 import Logger from '../../../common/lib/util/logger';
+import { StandardCallback } from '../../../common/types/utils';
+import { createRequest, Request } from '../transport/jsonptransport';
 
 function shouldFallback(errorInfo: ErrorInfo) {
 	const statusCode = errorInfo.statusCode as number;
@@ -38,6 +40,7 @@ const Http: typeof IHttp = class {
 	static methods = [HttpMethods.Get, HttpMethods.Delete, HttpMethods.Post, HttpMethods.Put, HttpMethods.Patch];
 	static methodsWithoutBody = [HttpMethods.Get, HttpMethods.Delete];
 	static methodsWithBody = [HttpMethods.Post, HttpMethods.Put, HttpMethods.Patch];
+	checksInProgress: Array<StandardCallback<boolean>> | null = null;
 
 	constructor() {
 		if(Platform.xhrSupported) {
@@ -56,6 +59,37 @@ const Http: typeof IHttp = class {
 					const result = (!err && (responseText as string)?.replace(/\n/, '') == 'yes');
 					Logger.logAction(Logger.LOG_MICRO, '(XHRRequest)Http.checkConnectivity()', 'Result: ' + result);
 					callback(null, result);
+				});
+			};
+		} else if (Platform.jsonpSupported) {
+			this.Request = function(method: HttpMethods, rest: Rest | null, uri: string, headers: Record<string, string> | null, params: RequestParams, body: unknown, callback: RequestCallback) {
+				const req = createRequest(uri, headers, params, body, XHRStates.REQ_SEND, rest && rest.options.timeouts, method);
+				req.once('complete', callback);
+				Utils.nextTick(function() {
+					req.exec();
+				});
+				return req;
+			};
+	
+			this.checkConnectivity = function(callback: (err: ErrorInfo | null, connectivity?: boolean) => void) {
+				const upUrl = Defaults.jsonpInternetUpUrl;
+	
+				if(this.checksInProgress) {
+					this.checksInProgress.push(callback);
+					return;
+				}
+				this.checksInProgress = [callback];
+				Logger.logAction(Logger.LOG_MICRO, '(JSONP)Http.checkConnectivity()', 'Sending; ' + upUrl);
+	
+				const req = new Request('isTheInternetUp', upUrl as string, null, null, null, XHRStates.REQ_SEND, Defaults.TIMEOUTS);
+				req.once('complete', (err: Error, response: string) => {
+					const result = !err && response;
+					Logger.logAction(Logger.LOG_MICRO, '(JSONP)Http.checkConnectivity()', 'Result: ' + result);
+					for(let i = 0; i < (this.checksInProgress as Array<StandardCallback<boolean>>).length; i++) (this.checksInProgress as Array<StandardCallback<boolean>>)[i](null, result);
+					this.checksInProgress = null;
+				});
+				Utils.nextTick(function() {
+					req.exec();
 				});
 			};
 		}
